+++ src/declaration/index.d.ts
@@ -145,3 +145,13 @@ export function removeEventListener<E extends EventTarget>(
   listener: EventListener | EventListenerObject,
   options?: boolean | EventListenerOptions
 ): void;
+
+/**
+ * Dispatches a synthetic event event to element and returns true if
+ * either event's cancelable attribute value is false or its preventDefault()
+ * method was not invoked, and false otherwise.
+ */
+export function dispatchEvent<E extends EventTarget>(
+  element: E,
+  event: Event
+): boolean;